# backend/openapi/schema.yml
openapi: 3.0.3
info:
  title: SOBASED API
  version: 1.0.0
paths:
...
  /api/public/share/{token}/:
    get:
      operationId: public_share_retrieve
      description: 'The page''s data: password gate, or the items with their media
        URLs.'
      parameters:
      - in: path
        name: token
        schema:
          type: string
        required: true
      tags:
      - public
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicShare'
          description: ''
        '404':
          description: No response body
        '410':
          description: No response body
...
components:
  schemas:
...
    PublicItem:
      type: object
      description: One playable / viewable thing of the public page.
      properties:
        version_id:
          type: integer
        name:
          type: string
        kind:
          type: string
        label:
          type: string
        number:
          type: integer
        mime_type:
          type: string
        duration_ms:
          type: integer
          nullable: true
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        page_count:
          type: integer
          nullable: true
        media_url:
          type: string
          nullable: true
        download_url:
          type: string
          nullable: true
        peaks_url:
          type: string
          nullable: true
        thumbnail_url:
          type: string
          nullable: true
        ready:
          type: boolean
          description: False while a watermark is being built
        error:
          type: string
      required:
      - download_url
      - duration_ms
      - error
      - height
      - kind
      - label
      - media_url
      - mime_type
      - name
      - number
      - page_count
      - peaks_url
      - ready
      - thumbnail_url
      - version_id
      - width
    PublicShare:
      type: object
      properties:
        title:
          type: string
        requires_password:
          type: boolean
        target_type:
          $ref: '#/components/schemas/ShareTargetEnum'
        allow_download:
          type: boolean
        watermark:
          type: boolean
        expires_at:
          type: string
          format: date-time
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/PublicItem'
      required:
      - allow_download
      - expires_at
      - items
      - requires_password
      - target_type
      - title
      - watermark
...
    ShareTargetEnum:
      enum:
      - version
      - asset
      - playlist
      type: string
      description: |-
        * `version` - Une version précise
        * `asset` - Un fichier (dernière version)
        * `playlist` - Une sélection de fichiers
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid